feat(daily_banking/sent_transfers): ✨ updated amounts in SenTransfersFilers DTO to be double

+++ lib/application/daily_banking/accounts/wires/sent_transfers/filter/filter_simplified_sent_transfers_controller.dart
@@ -39,11 +39,11 @@ class FilterSimplifiedSentTransfersController
     state = state.copyWith(endDate: endDate);
   }
 
-  void setAmountFrom(int? amountFrom) {
+  void setAmountFrom(double? amountFrom) {
     state = state.copyWith(amountFrom: amountFrom);
   }
 
-  void setAmountTo(int? amountTo) {
+  void setAmountTo(double? amountTo) {
     state = state.copyWith(amountTo: amountTo);
   }
 }
