+++ lib/infrastructure/daily_banking/wires/sent_transfers/repositories/sent_transfers_pagination_repository.dart
@@ -34,8 +34,8 @@ class SentTransfersPaginationRepository extends PaginationListRepository<Simplif
   }
 
   void updateFilter({
-    required int? amountFrom,
-    required int? amountTo,
+    required double? amountFrom,
+    required double? amountTo,
     required DateTime? startDateFrom,
     required DateTime? startDateTo,
   }) {
